Extract connected observation detail view
